+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Support for Cirrus template variables sourced from SSM parameters
+
 ### Changed
 
 ### Fixed
@@ -19,7 +21,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
--  Support for External AWS accounts Publish to Cirrus Proccess Queue.
+-  Support for External AWS accounts Publish to Cirrus Proccess Queue
 
 ## [2.50.0] - 2025-07-17